Reorginsation engine manifest upgrade code
